*** empty log message ***
git-svn-id: https://crosswire.org/svn/sword/trunk@282 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/modules/common/lzsscomprs.cpp
@@ -650,9 +650,9 @@ void LZSSCompress::Decode(void)
 				r = (short int) ( (r + 1) & (N - 1) );
 			}
 
-			// Add the "len" characters to the output stream.
+			// Add the "len" :characters to the output stream.
 
-			if (SendChars((char *) c, len) != len)
+			if (SendChars((char *) c, len) != (unsigned int)len)
 				break;
 		}
 	}
